pyproject.toml
[project]
name = "akshare-one"
version = "0.5.0"
description = "Standardized interface for Chinese financial market data, built on AKShare with unified data formats and simplified APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.17.80",
    "cachetools>=5.5.0",
]
license = "MIT"
keywords = ["akshare", "financial-data", "stock-data", "quant"]

[project.scripts]
akshare-one-mcp = "akshare_one_mcp.server:run_server"

[project.urls]
Homepage = "https://github.com/zwldarren/akshare-one"
Repository = "https://github.com/zwldarren/akshare-one.git"

[project.optional-dependencies]
talib = ["ta-lib>=0.6.4"]
mcp = [
    "fastmcp>=2.11.3",
    "pydantic>=2.0.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "mkdocs-material>=9.6.15",
    "pandas-stubs>=2.3.0.250703",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.1",
    "ty>=0.0.10",
    "types-cachetools>=6.1.0.20250717",
    "types-requests>=2.32.4.20250809",
    "fastmcp>=2.11.3",
    "pydantic>=2.0.0",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/akshare_one", "src/akshare_one_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=akshare_one --cov-report=term-missing"
markers = [
    "contract: Mark tests as contract tests",
    "integration: Mark tests as integration tests (require network)",
    "slow: Mark tests as slow running",
]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    "E501",  # Line too long (keep for readability)
    "E402",  # Module level import not at top (sometimes needed for optional imports)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "B018",  # Useless expression (common in tests)
    "SIM115",  # Use contextlib.suppress (not always better in tests)
]
"src/akshare_one/modules/*/__init__.py" = [
    "F401",  # Unused import (re-exports)
]

[tool.coverage.run]
source = ["src/akshare_one"]
omit = [
    "src/akshare_one/mcp/*",  # Optional MCP server
    "src/akshare_one/*/__init__.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 60
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
]
